train_img_re.py

+++ train_scrc.py
@@ -219,12 +219,12 @@ for i in range(2):
 
 model = utils.initialize_model(args.classifier,
                                num_classes=n_classes,
-                               chn_dim=len(scrc_in)).half()
+                               chn_dim=len(scrc_in))
 # for layer in model.modules():
 #   if isinstance(layer, torch.nn.BatchNorm2d):
 #     layer.float()
 # model.to(device)
-model.half()
+# model.half()
 optimizer = optim.Adam(model.parameters(), lr=args.lr, eps=1e-3)
 parameters = filter(lambda p: p.requires_grad, model.parameters())
 model, optimizer, _, __ = deepspeed.initialize(args=args,
@@ -258,7 +258,7 @@ for epoch in range(args.begin_epoch, args.nepochs):
         bat_id = np.random.rand(x.shape[0]).argsort()
         x = x[bat_id, ]
         y = y[bat_id, ]
-        x = x.half().to(device)
+        x = x.to(device)
         y = y.to(device)
 
         logits = model(x)
@@ -279,7 +279,7 @@ for epoch in range(args.begin_epoch, args.nepochs):
     model.eval()
     tot, cor = 0, 0
     for _, (x, y) in enumerate(tst_loader[1]):
-        x = x.half().to(device)
+        x = x.to(device)
         y = y.to(device)
         lgts = model(x)
         _, pred = lgts.max(1)
